changing design to remove console warnings
